requirements.txt
# requirements.txt
paddleocr>=2.7.0
paddlepaddle>=3.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
python-docx>=1.1.0
opencv-python>=4.8.0
pyspellchecker>=0.7.0
numpy>=1.24.0